requirements.txt
tqdm>=4.66.1
chromadb>=0.4.22
sentence-transformers
google-generativeai>=0.3.2
pathlib>=1.0.1
fastapi>=0.109.0
uvicorn>=0.27.0
